Update README.md
Hogwarts Four is a take on the classic game of Connect Four but with the user's being immersed with Hogwarts culture. At Hogwarts, each house fights for house points-- at then end of the year those points are counted up and the winning house obtains a great honor: The House Cup.